backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
opencv-python==4.11.0.86
mediapipe==0.10.5
celery==5.4.0
redis==5.0.0
pymongo==4.6.1
python-dotenv==1.0.1
Werkzeug>=3.1.0
numpy==1.26.4
pytz==2024.1
pytest==8.0.2
pytest-flask==1.3.0
gunicorn==22.0.0
